requirements.txt
blinker==1.7.0
certifi==2024.2.2
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.5
filelock==3.8.0
fire==0.4.0
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
identify==2.5.3
idna==3.6
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
nodeenv==1.7.0
numpy==1.26.4
opencv-python==4.9.0.80
openvino==2023.3.0
openvino-telemetry==2023.2.1
packaging==23.2
paho-mqtt==2.0.0
pandas==2.2.1
pillow==10.2.0
platformdirs==2.5.2
pre-commit==2.17.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
supervision==0.18.0
sympy==1.12
termcolor==1.1.0
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.18
urllib3==2.2.1
virtualenv==20.16.3
Werkzeug==3.0.1
